% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{ma_abstract}
\alias{ma_abstract}
\title{Abstracts}
\usage{
ma_abstract(query, key = NULL, ...)
}
\arguments{
\item{key}{(character) microsoft academic API key, see Details.}
}
\description{
Abstracts
}
\examples{
\dontrun{
ma_abstract(query = "Ti='biology'...")
}
}

