% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calchistogram.R
\name{ma_calchist}
\alias{ma_calchist}
\title{CalcHistogram API}
\usage{
ma_calchist()
}
\arguments{
\item{key}{(character) microsoft academic API key, see Details.}
}
\description{
CalcHistogram API
}
\examples{
\dontrun{
ma_calchist(query = "Ti='biology'...")
}
}

