% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{ma_evaluate}
\alias{ma_evaluate}
\title{Interpret API}
\usage{
ma_evaluate(query, count = 10, offset = 0, orderby = NULL,
  atts = c("Id", "AA.AuN", "J.JN", "Ti", "Y", "E", "CC"), key = NULL, ...)
}
\arguments{
\item{query}{(character) query terms}

\item{count}{(intger) number of records to return. default: 10}

\item{offset}{(intger) record number to start at. default: 0}

\item{orderby}{(logical) column by which to order. default: none}

\item{atts}{(character) fields to return, in a character vector. See
XX for details.}

\item{key}{(character) microsoft academic API key, see Details.}

\item{complete}{(logical) record to start at. default: 0}
}
\description{
Interpret API
}
\examples{
\dontrun{
ma_evaluate(query = "Ti='biology'...")
}
}

