% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{ma_interpret}
\alias{ma_interpret}
\title{Interpret API}
\usage{
ma_interpret(query, count = 10, complete = TRUE, key, ...)
}
\arguments{
\item{query}{(character) query terms}

\item{count}{(intger) number of records to return. default: 10}

\item{complete}{(logical) record to start at. default: 0}

\item{key}{(character) microsoft academic API key, see Details.}

\item{offset}{(intger) record number to start at. default: 0}
}
\description{
Interpret API
}
\examples{
\dontrun{
ma_interpret(query = "Ti='biology'...")
}
}

