% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ma_search}
\alias{ma_search}
\title{Search - higher level method}
\usage{
ma_search(
  query,
  count = 10,
  offset = 0,
  orderby = NULL,
  atts = c("Id", "AA.AuN", "J.JN", "Ti", "Y", "CC"),
  model = "latest",
  key = NULL,
  ...
)
}
\arguments{
\item{query}{(character) query terms}

\item{count}{(intger) number of records to return. default: 10}

\item{offset}{(intger) record number to start at. default: 0}

\item{orderby}{(logical) column by which to order. default: none}

\item{atts}{(character) fields to return, in a character vector. See
https://docs.microsoft.com/en-us/academic-services/project-academic-knowledge/reference-entity-attributes
for details.}

\item{model}{(character) Name of the model that you wish to query. One
of 'latest' or 'beta-2015'. Default: latest}

\item{key}{(character) microsoft academic API key, see the \code{Authentication}
section in \link{microdemic-package}}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Search - higher level method
}
\examples{
\dontrun{
ma_search(query = "Y='19'...")
}
}
\references{
https://academic.microsoft.com/
https://www.microsoft.com/en-us/research/project/academic-knowledge/
https://docs.microsoft.com/en-us/academic-services/project-academic-knowledge/introduction
}
