% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{ma_similarity}
\alias{ma_similarity}
\title{Similarity API}
\usage{
ma_similarity()
}
\arguments{
\item{key}{(character) microsoft academic API key, see Details.}
}
\description{
Similarity API
}
\examples{
\dontrun{
ma_similarity(query = "Ti='biology'...")
}
}

