% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdemic-package.R
\docType{package}
\name{microdemic-package}
\alias{microdemic}
\alias{microdemic-package}
\title{Microsoft Academic Client}
\description{
Microsoft Academic Client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

